+++ shared/Cargo.toml
@@ -7,5 +7,5 @@ edition = "2021"
 
 [dependencies]
 serde = { version = "1.0.147", features = ["derive"] }
-
+md5 = "0.7.0"
 serde_json = "1.0"
